Moved the extra factory up as non-extra module
